test(release-kit): split contract validation out, pin the decision matrix
Splits crates/landmark/src/release_kit.rs's self-contained JSON-schema
contract validation (assert_contract and its supporting validators,
~350 lines) into its own module, release_kit_contract.rs. This is
what backlog 012 asked for before adding tests: release_kit.rs was at
1012 lines under 011's deliberate, explicitly-reasoned ratchet bump to
1050; growing it further without splitting would have hit that cap
again immediately. release_kit.rs is now 666 lines and
bin/check-architecture's cap reverts to 1000 -- headroom restored, not
spent.

Adds crates/landmark/src/release_kit_tests.rs (mirroring the existing
version_decision_tests.rs pattern) with direct unit tests for
release_kit_importance (every branch: security, migration via major
bump/breaking/migration_heavy, high, launch, low, and the medium
default -- plus the launch/medium boundary case where a bootstrap
release has no bump signal), release_kit_needs_rich_artifacts (every
importance value), and release_kit_audiences (release-operator/
docs-owner gated correctly on rich-artifact need, primary + developer-
operator always present). These were bare private fn; bumped to
pub(crate) so the sibling test file can reach them, matching how
version_decision.rs's classify_commit/decide_version are exposed for
the same reason.

Backlog: 012
