+++ bin/check-architecture
@@ -21,12 +21,7 @@ check_max_lines() {
 }
 
 check_max_lines "crates/landmark/src/main.rs" 120
-# release_kit.rs grew past 1000 landing backlog 011 (structured-commit
-# classification, closing the second bare-text classifier misfire site).
-# Small, deliberate headroom bump, not a blank check: backlog 012 adds direct
-# unit tests to this file next and will need to split ownership (plan/build
-# vs. assert_contract validation) rather than raise this cap again.
-check_max_lines "crates/landmark/src/release_kit.rs" 1050
+check_max_lines "crates/landmark/src/release_kit.rs" 1000
 
 while IFS= read -r path; do
   case "${path}" in
